+++ modules/MSP365.Reporting/ChangeLog.md
@@ -4,3 +4,4 @@
 - **1.0.14** Added the Get-M365Report function
 - **1.0.15** Added the Get-M365Report function
 - **1.0.16** Added the Get-M365Report function
+- **1.0.17** Added the Get-M365Report function