+++ modules/MSP365.SAM/ChangeLog.md
@@ -3,3 +3,4 @@
 - **1.0.0** Created the module
 - **1.0.11** Added the Get-M365SAMStatus function
 - **1.0.12** Added the Get-M365Report function
+- **1.0.13** Added the Get-M365Report function