+++ modules/MSP365/ChangeLog.md
@@ -7,3 +7,4 @@
 - **1.0.15** Updated Reporting and SAM modules
 - **1.0.16** Updated Reporting and SAM modules
 - **1.0.17** Updated Reporting and SAM modules
+- **1.0.18** Updated Reporting and SAM modules